+++ shared/libebm/compute/BinSumsBoosting.hpp
@@ -55,22 +55,22 @@ GPU_DEVICE NEVER_INLINE static void BinSumsBoostingInternal(BinSumsBoostingBridg
    if(!bCompilerZeroDimensional) {
       cBytesPerBin = static_cast<typename TFloat::TInt::T>(GetBinSize<typename TFloat::T, typename TFloat::TInt::T>(bHessian, cScores));
 
-      const ptrdiff_t cPack = GET_ITEMS_PER_BIT_PACK(cCompilerPack, pParams->m_cPack);
+      const int cPack = GET_ITEMS_PER_BIT_PACK(cCompilerPack, pParams->m_cPack);
 #ifndef GPU_COMPILE
-      EBM_ASSERT(k_cItemsPerBitPackNone != cPack); // we require this condition to be templated
+      EBM_ASSERT(static_cast<int>(k_cItemsPerBitPackNone) != cPack); // we require this condition to be templated
       EBM_ASSERT(1 <= cPack);
 #endif // GPU_COMPILE
 
-      const int cItemsPerBitPack = static_cast<int>(cPack);
+      const int cItemsPerBitPack = cPack;
 #ifndef GPU_COMPILE
       EBM_ASSERT(1 <= cItemsPerBitPack);
-      EBM_ASSERT(static_cast<size_t>(cItemsPerBitPack) <= COUNT_BITS(typename TFloat::TInt::T));
+      EBM_ASSERT(cItemsPerBitPack <= COUNT_BITS(typename TFloat::TInt::T));
 #endif // GPU_COMPILE
 
-      cBitsPerItemMax = static_cast<int>(GetCountBits<typename TFloat::TInt::T>(static_cast<size_t>(cItemsPerBitPack)));
+      cBitsPerItemMax = GetCountBits<typename TFloat::TInt::T>(cItemsPerBitPack);
 #ifndef GPU_COMPILE
       EBM_ASSERT(1 <= cBitsPerItemMax);
-      EBM_ASSERT(static_cast<size_t>(cBitsPerItemMax) <= COUNT_BITS(typename TFloat::TInt::T));
+      EBM_ASSERT(cBitsPerItemMax <= COUNT_BITS(typename TFloat::TInt::T));
 #endif // GPU_COMPILE
 
       cShift = static_cast<int>(((cSamples >> TFloat::k_cSIMDShift) - size_t { 1 }) % static_cast<size_t>(cItemsPerBitPack)) * cBitsPerItemMax;
@@ -336,7 +336,7 @@ INLINE_RELEASE_TEMPLATED ErrorEbm OperatorBinSumsBoosting(BinSumsBoostingBridge 
 
 template<typename TFloat, bool bHessian, size_t cCompilerScores, bool bWeight, bool bReplication>
 INLINE_RELEASE_TEMPLATED static ErrorEbm BitPackBoosting(BinSumsBoostingBridge * const pParams) {
-   if(k_cItemsPerBitPackNone != pParams->m_cPack) {
+   if(static_cast<int>(k_cItemsPerBitPackNone) != pParams->m_cPack) {
       return OperatorBinSumsBoosting<TFloat, bHessian, cCompilerScores, bWeight, bReplication, k_cItemsPerBitPackDynamic>(pParams);
    } else {
       // this needs to be special cased because otherwise we would inject comparisons into the dynamic version
